update io signature

+++ emp-tool/io/io_channel.h
@@ -10,25 +10,25 @@ namespace emp {
 template<typename T> 
 class IOChannel { public:
 	uint64_t counter = 0;
-	void send_data(const void * data, int nbyte) {
+	void send_data(const void * data, size_t nbyte) {
 		counter +=nbyte;
 		derived().send_data_internal(data, nbyte);
 	}
 
-	void recv_data(void * data, int nbyte) {
+	void recv_data(void * data, size_t nbyte) {
 		derived().recv_data_internal(data, nbyte);
 	}
 
-	void send_block(const block* data, int nblock) {
+	void send_block(const block* data, size_t nblock) {
 		send_data(data, nblock*sizeof(block));
 	}
 
-	void recv_block(block* data, int nblock) {
+	void recv_block(block* data, size_t nblock) {
 		recv_data(data, nblock*sizeof(block));
 	}
 
-	void send_pt(Point *A, int num_pts = 1) {
-		for(int i = 0; i < num_pts; ++i) {
+	void send_pt(Point *A, size_t num_pts = 1) {
+		for(size_t i = 0; i < num_pts; ++i) {
 			size_t len = A[i].size();
 			A[i].group->resize_scratch(len);
 			unsigned char * tmp = A[i].group->scratch;
@@ -38,9 +38,9 @@ class IOChannel { public:
 		}
 	}
 
-	void recv_pt(Group * g, Point *A, int num_pts = 1) {
+	void recv_pt(Group * g, Point *A, size_t num_pts = 1) {
 		size_t len = 0;
-		for(int i = 0; i < num_pts; ++i) {
+		for(size_t i = 0; i < num_pts; ++i) {
 			recv_data(&len, 4);
 			assert(len <= 2048);
 			g->resize_scratch(len);
@@ -50,36 +50,36 @@ class IOChannel { public:
 		}
 	}	
 
-	void send_bool(bool * data, int length) {
+	void send_bool(bool * data, size_t length) {
 		void * ptr = (void *)data;
 		size_t space = length;
 		const void * aligned = std::align(alignof(uint64_t), sizeof(uint64_t), ptr, space);
 		if(aligned == nullptr)
 			send_data(data, length);
 		else{
-			int diff = length - space;
+			size_t diff = length - space;
 			send_data(data, diff);
 			send_bool_aligned((const bool*)aligned, length - diff);
 		}
 	}
 
-	void recv_bool(bool * data, int length) {
+	void recv_bool(bool * data, size_t length) {
 		void * ptr = (void *)data;
 		size_t space = length;
 		void * aligned = std::align(alignof(uint64_t), sizeof(uint64_t), ptr, space);
 		if(aligned == nullptr)
 			recv_data(data, length);
 		else{
-			int diff = length - space;
+			size_t diff = length - space;
 			recv_data(data, diff);
 			recv_bool_aligned((bool*)aligned, length - diff);
 		}
 	}
 
 
-	void send_bool_aligned(const bool * data, int length) {
+	void send_bool_aligned(const bool * data, size_t length) {
 		unsigned long long * data64 = (unsigned long long * )data;
-		int i = 0;
+		size_t i = 0;
 		for(; i < length/8; ++i) {
 			unsigned long long mask = 0x0101010101010101ULL;
 			unsigned long long tmp = 0;
@@ -97,9 +97,9 @@ class IOChannel { public:
 		if (8*i != length)
 			send_data(data + 8*i, length - 8*i);
 	}
-	void recv_bool_aligned(bool * data, int length) {
+	void recv_bool_aligned(bool * data, size_t length) {
 		unsigned long long * data64 = (unsigned long long *) data;
-		int i = 0;
+		size_t i = 0;
 		for(; i < length/8; ++i) {
 			unsigned long long mask = 0x0101010101010101ULL;
 			unsigned long long tmp = 0;
